A Function for Descriptive data for item analysis.

This function will create a publication ready essential descriptive table for item analysis. Normality is tested using shapiro.test from base stats with Bonferroni Correction.

Usage

des.tab(df, reverse = FALSE)

Arguments

df

A data frame.

reverse

If TRUE, will provide indicate which items had a negative correlation and reverse them

Value

Returns a summary table of descriptives in a data frame structure.

Examples

data <- tabledown::Rotter[, 11:31]
table <- des.tab(data)