Produce Publication Quality Tables and Plots

The tabledown package provides necessary data frames used throughout the book and some neat functions.

tabledown Data-frames

Rotter: Psychometric validation data of Bangla Rotter's Internal- External Scale.

Gantt and Spot: Two sample data-frames for creating project management Gantt chart.

FFMQ.CFA: Structural Validation data of Bangla Five Factor Mindfulness Questionnaire.

FFMQ.Val:Correlational Validity evidences of Bangla Five Factor Mindfulness Questionnaire.

tabledown functions

This packages includes some neat and useful functions to create tables and figures suitable for journal submission:

fac.tab(): Creates a publication ready table from the output of "psych" package based factor analysis.

des.tab(): Creates a publication ready descriptive table of Item analysis with the reporting of normality assumptions.

normality.loop(): Compute normality test on the whole data frame. No grouping variable required.

bedTime(): Calculate total time spent in bed from the sleep log entry.

cfa.tab():Creates a table with necessary fit indices from a "lavaan" class objects.

cfa.tab/multi():creates a table with necessary fit indices from several lavaan class objects.

ggicc: Creates a ggplot2 based publication ready Item Characteristics Curve from the "mirt" package based item response theory estimations.

ggiteminfo: Creates a ggplot2 based publication ready Item Information Curve from the "mirt" package based item response theory estimations.

ggtestinfo: Creates a ggplot2 based publication ready Test Information Curve from the "mirt" package based item response theory estimations.

ggtestinfo_se: Creates a ggplot2 based publication ready Test Information Curve with standard error from the "mirt" package based item response theory estimations. It is advisable that you load tidyverse along with tabledown